renaming db_common to common
this moves the following methods into common
  arg_host_range ( used in creds and db )
  arg_port_range ( used in creds and db )
  set_rhosts_from_addrs ( used in creds and db )
  show_options ( used in jobs and modules )
