+++ lib/msf/ui/console/command_dispatcher/creds.rb
@@ -3,7 +3,6 @@
 require 'rexml/document'
 require 'rex/parser/nmap_xml'
 require 'msf/core/db_export'
-require 'msf/ui/console/command_dispatcher/db_common'
 
 module Msf
 module Ui
@@ -15,7 +14,7 @@ class Creds
 
   include Msf::Ui::Console::CommandDispatcher
   include Metasploit::Credential::Creation
-  include Msf::Ui::Console::CommandDispatcher::DbCommon
+  include Msf::Ui::Console::CommandDispatcher::Common
   
   #
   # The dispatcher's name.
@@ -54,39 +53,6 @@ class Creds
   end
   
   #
-  # Miscellaneous option helpers
-  #
-
-  # Parse +arg+ into a {Rex::Socket::RangeWalker} and append the result into +host_ranges+
-  #
-  # @note This modifies +host_ranges+ in place
-  #
-  # @param arg [String] The thing to turn into a RangeWalker
-  # @param host_ranges [Array] The array of ranges to append
-  # @param required [Boolean] Whether an empty +arg+ should be an error
-  # @return [Boolean] true if parsing was successful or false otherwise
-  def arg_host_range(arg, host_ranges, required=false)
-    if (!arg and required)
-      print_error("Missing required host argument")
-      return false
-    end
-    begin
-      rw = Rex::Socket::RangeWalker.new(arg)
-    rescue
-      print_error("Invalid host parameter, #{arg}.")
-      return false
-    end
-
-    if rw.valid?
-      host_ranges << rw
-    else
-      print_error("Invalid host parameter, #{arg}.")
-      return false
-    end
-    return true
-  end
-  
-  #
   # Can return return active or all, on a certain host or range, on a
   # certain port or range, and/or on a service name.
   #
@@ -118,6 +84,9 @@ class Creds
   # TODO: this needs to be cleaned up to use the new syntax
   #
   def cmd_creds_help
+    require 'pry'
+    binding.pry
+    
     print_line
     print_line "With no sub-command, list credentials. If an address range is"
     print_line "given, show only credentials with logins on hosts within that"
